Hide the navbar on the landing page
I made the decision to hide the navbar on the landing page as there are
only 2 links and it is not necesary. We can descibe the app and include
the action items (sign up or sign in) as buttons on the landing. Navbar
can be added back in once there are additional items in the navigation
if necessary.